# docs/swagger.yaml
basePath: /api/v1
definitions:
  github_com_EduGoGroup_edugo-api-admin-new_internal_application_dto.AcademicUnitResponse:
    properties:
      code:
        type: string
      created_at:
        type: string
      deleted_at:
        type: string
      description:
        type: string
      display_name:
        type: string
      id:
        type: string
      metadata:
        additionalProperties: true
        type: object
      parent_unit_id:
        type: string
      school_id:
        type: string
      type:
        type: string
      updated_at:
        type: string
    type: object
  github_com_EduGoGroup_edugo-api-admin-new_internal_application_dto.CreateAcademicUnitRequest:
    properties:
      code:
        type: string
      description:
        type: string
      display_name:
        minLength: 3
        type: string
      metadata:
        additionalProperties: true
        type: object
      parent_unit_id:
        type: string
      type:
        type: string
    required:
    - display_name
    - type
    type: object
...
  github_com_EduGoGroup_edugo-api-admin-new_internal_application_dto.ErrorResponse:
    properties:
      code:
        type: string
      details:
        additionalProperties:
          type: string
        type: object
      error:
        type: string
    type: object
...
  github_com_EduGoGroup_edugo-api-admin-new_internal_application_dto.UpdateAcademicUnitRequest:
    properties:
      description:
        type: string
      display_name:
        type: string
      metadata:
        additionalProperties: true
        type: object
      parent_unit_id:
        type: string
    type: object
...
host: localhost:8081
info:
  contact: {}
  description: Clean administration API for EduGo
  title: EduGo API Admin New
  version: "1.0"
paths:
...
  /schools/{id}/units:
    get:
      consumes:
      - application/json
      parameters:
      - description: School ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_EduGoGroup_edugo-api-admin-new_internal_application_dto.AcademicUnitResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_EduGoGroup_edugo-api-admin-new_internal_application_dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_EduGoGroup_edugo-api-admin-new_internal_application_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_EduGoGroup_edugo-api-admin-new_internal_application_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List academic units by school
      tags:
      - academic-units
    post:
      consumes:
      - application/json
      parameters:
      - description: School ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Academic unit data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_EduGoGroup_edugo-api-admin-new_internal_application_dto.CreateAcademicUnitRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_EduGoGroup_edugo-api-admin-new_internal_application_dto.AcademicUnitResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_EduGoGroup_edugo-api-admin-new_internal_application_dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_EduGoGroup_edugo-api-admin-new_internal_application_dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_EduGoGroup_edugo-api-admin-new_internal_application_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_EduGoGroup_edugo-api-admin-new_internal_application_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create an academic unit under a school
      tags:
      - academic-units
  /schools/{id}/units/by-type:
    get:
      consumes:
      - application/json
      parameters:
      - description: School ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Unit type filter
        in: query
        name: type
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_EduGoGroup_edugo-api-admin-new_internal_application_dto.AcademicUnitResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_EduGoGroup_edugo-api-admin-new_internal_application_dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_EduGoGroup_edugo-api-admin-new_internal_application_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_EduGoGroup_edugo-api-admin-new_internal_application_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List academic units by type for a school
      tags:
      - academic-units
  /schools/{id}/units/tree:
    get:
      consumes:
      - application/json
      parameters:
      - description: School ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_EduGoGroup_edugo-api-admin-new_internal_application_dto.AcademicUnitResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_EduGoGroup_edugo-api-admin-new_internal_application_dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_EduGoGroup_edugo-api-admin-new_internal_application_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_EduGoGroup_edugo-api-admin-new_internal_application_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get hierarchical unit tree for a school
      tags:
      - academic-units
...
  /units/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Academic Unit ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_EduGoGroup_edugo-api-admin-new_internal_application_dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_EduGoGroup_edugo-api-admin-new_internal_application_dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_EduGoGroup_edugo-api-admin-new_internal_application_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_EduGoGroup_edugo-api-admin-new_internal_application_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Soft delete an academic unit
      tags:
      - academic-units
    get:
      consumes:
      - application/json
      parameters:
      - description: Academic Unit ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_EduGoGroup_edugo-api-admin-new_internal_application_dto.AcademicUnitResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_EduGoGroup_edugo-api-admin-new_internal_application_dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_EduGoGroup_edugo-api-admin-new_internal_application_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_EduGoGroup_edugo-api-admin-new_internal_application_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get an academic unit by ID
      tags:
      - academic-units
    put:
      consumes:
      - application/json
      parameters:
      - description: Academic Unit ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Academic unit update data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_EduGoGroup_edugo-api-admin-new_internal_application_dto.UpdateAcademicUnitRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_EduGoGroup_edugo-api-admin-new_internal_application_dto.AcademicUnitResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_EduGoGroup_edugo-api-admin-new_internal_application_dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_EduGoGroup_edugo-api-admin-new_internal_application_dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_EduGoGroup_edugo-api-admin-new_internal_application_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_EduGoGroup_edugo-api-admin-new_internal_application_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update an academic unit
      tags:
      - academic-units
  /units/{id}/hierarchy-path:
    get:
      consumes:
      - application/json
      parameters:
      - description: Academic Unit ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_EduGoGroup_edugo-api-admin-new_internal_application_dto.AcademicUnitResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_EduGoGroup_edugo-api-admin-new_internal_application_dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_EduGoGroup_edugo-api-admin-new_internal_application_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_EduGoGroup_edugo-api-admin-new_internal_application_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get hierarchy path from root to unit
      tags:
      - academic-units
  /units/{id}/restore:
    post:
      consumes:
      - application/json
      parameters:
      - description: Academic Unit ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_EduGoGroup_edugo-api-admin-new_internal_application_dto.AcademicUnitResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_EduGoGroup_edugo-api-admin-new_internal_application_dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_EduGoGroup_edugo-api-admin-new_internal_application_dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_EduGoGroup_edugo-api-admin-new_internal_application_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_EduGoGroup_edugo-api-admin-new_internal_application_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Restore a soft-deleted academic unit
      tags:
      - academic-units
...
securityDefinitions:
  BearerAuth:
    description: 'Type "Bearer" followed by a space and the JWT token. Example: "Bearer
      eyJhbGci..."'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"